+++ deepstate/with-sveltekit/src/routes/+page.svelte
@@ -1,21 +1,20 @@
 <script>
-  import { reify } from '@m5nv/deepstate/core';
-
-  const store = reify({ count: 0 }, { double: (state) => state.count * 2 });
-  const { _count, _double } = store.state;
-  const { state } = store;
+  const state = {
+    count: 0,
+    double: (state) => state.count * 2,
+  };
 </script>
 
-<h1>Welcome to SvelteKit</h1>
-<input type="number" value={state.count} />
+<h2>Plain Svelte Reactivity (no runes)</h2>
+<input type="number" bind:value={state.count} />
 <button onclick={() => (state.count += 1)}>count</button>
 <h3>Is it working?</h3>
-<code>{state.count} * 2 = {state.double}</code>
+<code>{state.count} * 2 = {state.double(state)}</code>
 <br />
-<code>{_count.value} * 2 = {_double.value}</code>
 
 <p>
-  Visit <a href="https://www.npmjs.com/package/@m5nv/deepstate"
-    >@m5nv/deepstate</a
-  > to read the documentation
+  Svelte provides fine grained reactivity by itself, even without having to use
+  runes. In this example, we can see that a plain object can be used without
+  much fuss to obtain "reactivity"; the only bit of twiddling we had to do was
+  to call the computed function instead of accessing it as a variable.
 </p>